Update ROCm 10 HIPIFY compatibility sources
Refresh the in-tree HIPIFY transformations and the ROCm-specific compatibility
branches across ATen, c10, distributed, AOTInductor, profiling and generated
extension code, so the gfx1201 build maps CUDA runtime and driver interfaces
onto their HIP equivalents and accounts for ROCm 10's API and status-enum
differences. Third-party submodule worktrees and untracked generated build
outputs are intentionally excluded.

ROCm 10 also breaks several late torch_hip targets, and those are fixed at
their source boundaries with the CUDA paths left intact: HIP and RCCL headers
leak __noinline__ into GCC 16's C++26 standard headers, current hipCUB wants
narrower component includes and rocPRIM reduction operators, and the RCCL and
MIOpen APIs have moved past older PyTorch call sites. Deprecation diagnostics
are suppressed only while the ROCm runtime function table is populated, since
ROCm 10 marks context entry points deprecated that PyTorch still resolves
dynamically.

What this deliberately does not do is commit the output of
tools/amd_build/build_amd.py. That script rewrites the tree in place, and its
substitutions do not port anything: they turn every __CUDACC__ into __HIPCC__,
every CUDA header include into its HIP counterpart and every CUDA type name
into the HIP one, including in the conditions that already covered both
compilers -- which deletes the CUDA half rather than adding a ROCm one. The
result does not build under nvcc at all: complex.h stops including
<thrust/complex.h> while c10/util/complex.h still calls thrust::abs, and
BFloat16.h and Half.h ask for <hip/hip_bf16.h> and <hip/hip_fp16.h> behind

One condition in torch/csrc/stable/c/shim.h is exactly the damage described
above and is spelled back: the guard around the CUDA stable shim's cuBLAS-handle
and c10 error-checking declarations read USE_CUDA, and rewriting it to USE_ROCM
removes those declarations from every CUDA build rather than adding a ROCm arm.

The same damage in torch/csrc/Module.cpp and the static launcher's cuda.{cpp,h}
was live for two weeks before it was caught: every USE_CUDA guard in Module.cpp
rewritten to USE_ROCM meant CUDA builds omitted their Python bindings, reported
_has_cuda as false, and could not compile or initialize the static launcher.
The canonical CUDA guards, the SDPA include, the Python module initialization
and the Driver API types are restored here, while the ROCm-specific branches and
the deprecated HIP context warning suppression stay.

CUDA spellings are the state this repository holds. A ROCm build runs
build_amd.py itself (.ci/pytorch/build.sh and the manywheel scripts) and
generates the HIP tree from them; .ci/pytorch/common_utils.sh even carries a
TODO asking for an option to revert the tree afterwards. HIP spellings inside
USE_ROCM and __HIPCC__ blocks stay -- CUDABlas.cpp's rocblas_status to
HIPBLAS_STATUS converter is real ROCm code, not damage.

Test Plan:

```
.venv/bin/python tools/amd_build/build_amd.py
python3 agent_space/dehipify.py        # re-detect: 0 files, 0 tokens
git diff --check
.venv/bin/lintrunner -a
```

```
source .venv/rocm-env.sh && CC=/usr/bin/clang CXX=/usr/bin/clang++ CMAKE_ARGS="-DCMAKE_C_COMPILER=/usr/bin/clang -DCMAKE_CXX_COMPILER=/usr/bin/clang++" USE_ROCM=1 USE_CUDA=0 USE_XPU=0 PYTORCH_ROCM_ARCH=gfx1201 BUILD_TEST=0 MAX_JOBS=16 USE_KINETO=0 USE_MSLK=0 USE_NVSHMEM=0 USE_HIPSPARSELT=0 USE_FLASH_ATTENTION=1 USE_MEM_EFF_ATTENTION=1 USE_ROCM_CK_GEMM=1 USE_ROCM_CK_SDPA=1 USE_ROCM_CK_SDPA_GFX12=1 .venv/bin/python -m pip install -e . -v --no-build-isolation
```

The re-detection pass reports no remaining hipified tokens, and it was run over
the whole tree rather than the directories already known to be broken -- an
earlier pass looked only at torch/headeronly and so missed everything below it.
Divergence from upstream over the 83 files the token pass touched fell from
3367 to 2304 diff lines, with no file moving further away. The incremental ROCm
build reached 773/1283 before the subsequent libhipcxx change required
namespace qualification; the final rebuild was interrupted for a rebase, so it
is not a completed build.

Authored with the assistance of an AI agent.

+++ aten/src/ATen/miopen/Descriptors.cpp
@@ -107,7 +107,8 @@ void TensorDescriptor::print() { std::cout << *this; }
 
 void FilterDescriptor::set(const at::Tensor &t, const at::MemoryFormat memory_format, int64_t pad) {
   auto dim = t.ndimension();
-  if (dim > MIOPEN_DIM_MAX || pad > MIOPEN_DIM_MAX)
+  if (dim > static_cast<int64_t>(MIOPEN_DIM_MAX) ||
+      pad > static_cast<int64_t>(MIOPEN_DIM_MAX))
   TORCH_CHECK(false, "MIOpen supports only up to ", MIOPEN_DIM_MAX, " dimensions");
   // NB: It is possible for this test to be insufficient, because the
   // Tensor passed in to set the filter descriptor may not be the actual
